refactor(formation): Standardize formation side effect handler and extract utilities
